adding corner cases

+++ controllers/fileOp.js
@@ -8,24 +8,28 @@ const FileCrypt = require("file-aes-crypt");
 exports.getEnc = async (req, res, next) => {
     const fileName = req.params.fileName;
     const FilePath = path.join('./public', `./uploads/${fileName}`);
-    const storePath = path.join('./public', `./uploads/${fileName}-enc`);
+    const storePath = path.join('./public', `./uploads/${fileName.replace('-dec', '')}-enc`);
     try {
-        const fc = new FileCrypt("fage");
-        let result = await fc.encrypt(FilePath, storePath);
-        console.log("encrypt ok");
-        deleteFile(FilePath);
-        // result = await fc.decrypt("./tmp/package-aes.json", "./temp/package2.json");
-        // console.log("decrypt ok");
+        if(!fileName.endsWith('-enc')){
+            
+            const fc = new FileCrypt("fage");
+            let result = await fc.encrypt(FilePath, storePath);
+            console.log("encrypt ok");
+            deleteFile(FilePath);
+            return res.redirect('/');
+        }
+        console.log('File is already encrypted');
+        res.redirect('/');
       } catch (e) {
         console.log(e);
       }
-    res.redirect('/');
+    
 }
 
 exports.getDec = async(req, res, next) => {
     const fileName = req.params.fileName;
     const FilePath = path.join('./public', `./uploads/${fileName}`);
-    const storePath = path.join('./public', `./uploads/${fileName}-dec`);
+    const storePath = path.join('./public', `./uploads/${fileName.replace('-enc','')}-dec.pdf`);
     try {
         const fc = new FileCrypt("fage");
         let result = await fc.decrypt(FilePath, storePath);
